pyproject.toml
[project]
name = "ag-ui-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.6",
    "langchain-openai>=1.1.8",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.2.0",
    "tavily-python>=0.7.21",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
